% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eezs}
\alias{eezs}
\title{Exclusive Economic Zones (EEZs)}
\format{
A SF object with the following attributes:
\describe{
  \item{eez_id}{EEZ id}
  \item{eez}{EEZ}
}
}
\source{
Flanders Marine Institute (2019) Maritime Boundaries Geodatabase, version 11. Available from: https://www.marineregions.org/
}
\usage{
eezs
}
\description{
Exclusive Economic Zones (EEZs) v11 shapefile from the Flanders Marine Institute.
}
\keyword{datasets}
