% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eezs_lines}
\alias{eezs_lines}
\title{Exclusive Economic Zones (EEZs) - boundaries}
\format{
A SF object with the following attributes:
\describe{
  \item{eez_id}{EEZ id}
  \item{eez}{EEZ}
}
}
\source{
Flanders Marine Institute (2019) Maritime Boundaries Geodatabase, version 11. Available from: https://www.marineregions.org/
}
\usage{
eezs_lines
}
\description{
Exclusive Economic Zones (EEZs) v11 boundaries shapefile from the Flanders Marine Institute.
}
\keyword{datasets}
