% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lmes}
\alias{lmes}
\title{Large Marine Ecoregions (LMEs)}
\format{
A SF object with the following attributes:
\describe{
  \item{ecoregion_id}{Ecoregion id}
  \item{ecoregion}{Ecoregion}
}
}
\source{
IUCN (2021) Large Marine Ecosystems of the World. Previously available at: http://lme.edc.uri.edu/index.php. Now potentially available at: https://lmehub.net/
}
\usage{
lmes
}
\description{
Large Marine Ecoregions (LMEs) shapefile.
}
\keyword{datasets}
