% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lmes}
\alias{lmes}
\title{Large Marine Ecoregions (LMEs)}
\format{
A SF object with the following attributes::
\describe{
  \item{lme_id}{}
  \item{lme_name}{}
}
}
\source{
https://lmehub.net/ - Spalding LME
}
\usage{
lmes
}
\description{
Large Marine Ecoregions (LMEs) shapefile.
}
\keyword{datasets}
