% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meows}
\alias{meows}
\title{Marine Ecoregions of the World (MEOWs)}
\format{
A SF object with the following attributes::
\describe{
  \item{lat_zone}{Latitude zone}
  \item{realm_id}{Realm id}
  \item{realm}{Realm}
  \item{province_id}{Province id}
  \item{province}Province{}
  \item{ecoregion_id}{Ecoregion id}
  \item{ecoregion}{Ecoregion}
}
}
\source{
https://lmehub.net/ - Spalding LME
}
\usage{
meows
}
\description{
Marine Ecoregions of the World (MEOWs) shapefile.
}
\keyword{datasets}
